# specs/backend/backend-api.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "System Design Simulator — Architecture Library API",
    "version": "1.0.0-draft",
    "description": "Release 1 design contract. Architecture CRUD and explicit Save accept incomplete diagrams. Simulation calculates the current unsaved document and records all forward requests and returning responses. The frontend owns unsaved edits. x-strict-integer and x-finite describe required runtime checks beyond standard JSON Schema."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Example local development address; configurable during implementation."
    }
  ],
...
  "paths": {
    "/api/v1/architectures": {
      "post": {
        "tags": [
          "Architectures"
        ],
        "operationId": "createArchitecture",
        "summary": "Create and save an architecture",
        "description": "First Save creates a new persisted architecture. Accepts incomplete diagrams. Every successful POST creates a new ID; retries are not idempotent.",
        "requestBody": {
          "$ref": "#/components/requestBodies/ArchitectureWrite"
        },
        "responses": {
          "201": {
            "description": "Architecture durably created. Creation and update timestamps are equal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Architecture"
                },
                "example": {
                  "id": "0ccdc365-31f4-46c8-9f94-2b36d7f73b08",
                  "name": "Application with database",
                  "created_at": "2026-10-03T15:00:00.000000Z",
                  "updated_at": "2026-10-03T15:00:00.000000Z",
                  "document": {
                    "format_version": 1,
                    "nodes": [
                      {
                        "id": "callers",
                        "type": "caller_group",
                        "label": "Callers",
                        "position": {
                          "x": -80,
                          "y": 120.5
                        },
                        "capacity_rps": null,
                        "caller_count": 100,
                        "test_rps": 100
                      },
                      {
                        "id": "lb",
                        "type": "load_balancer",
                        "label": "Load balancer",
                        "position": {
                          "x": 160,
                          "y": 120.5
                        },
                        "capacity_rps": 100,
                        "routing_policy": "weighted"
                      },
                      {
                        "id": "gateway",
                        "type": "gateway",
                        "label": "Gateway",
                        "position": {
                          "x": 400,
                          "y": 120.5
                        },
                        "capacity_rps": 100,
                        "routing_policy": "round_robin"
                      },
                      {
                        "id": "server",
                        "type": "server",
                        "label": "Server",
                        "position": {
                          "x": 640,
                          "y": 120.5
                        },
                        "capacity_rps": 60
                      },
                      {
                        "id": "database",
                        "type": "database",
                        "label": "Database",
                        "position": {
                          "x": 880,
                          "y": 120.5
                        },
                        "capacity_rps": 60
                      }
                    ],
                    "edges": [
                      {
                        "id": "callers-lb",
                        "source": "callers",
                        "target": "lb",
                        "order": 0,
                        "weight": 1
                      },
                      {
                        "id": "lb-gateway",
                        "source": "lb",
                        "target": "gateway",
                        "order": 0,
                        "weight": 1.5
                      },
                      {
                        "id": "gateway-server",
                        "source": "gateway",
                        "target": "server",
                        "order": 0,
                        "weight": 0
                      },
                      {
                        "id": "server-database",
                        "source": "server",
                        "target": "database",
                        "order": 0,
                        "weight": 1
                      }
                    ]
                  }
                }
              }
            },
            "headers": {
              "Location": {
                "required": true,
                "description": "Relative route for the newly saved architecture.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                },
                "example": "/api/v1/architectures/0ccdc365-31f4-46c8-9f94-2b36d7f73b08"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidJson"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "503": {
            "$ref": "#/components/responses/StorageUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "tags": [
          "Architectures"
        ],
        "operationId": "listArchitectures",
        "summary": "List saved architectures",
        "description": "Metadata only. All entries, ordered by updated_at descending then ID ascending. No pagination. An empty library returns items: [].",
        "responses": {
          "200": {
            "description": "Saved architecture metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchitectureList"
                },
                "examples": {
                  "library": {
                    "value": {
                      "items": [
                        {
                          "id": "0ccdc365-31f4-46c8-9f94-2b36d7f73b08",
                          "name": "Application with database",
                          "created_at": "2026-10-03T15:00:00.000000Z",
                          "updated_at": "2026-10-03T15:00:00.000000Z",
                          "format_version": 1
                        }
                      ]
                    }
                  },
                  "empty": {
                    "value": {
                      "items": []
                    }
                  }
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/StorageUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/architectures/{architecture_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ArchitectureId"
        }
      ],
      "get": {
        "tags": [
          "Architectures"
        ],
        "operationId": "getArchitecture",
        "summary": "Open a saved architecture",
        "description": "Return the latest saved document. Unknown well-formed UUIDs return 404; invalid UUID strings return 422.",
        "responses": {
          "200": {
            "description": "Complete latest saved architecture.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Architecture"
                },
                "example": {
                  "id": "0ccdc365-31f4-46c8-9f94-2b36d7f73b08",
                  "name": "Application with database",
                  "created_at": "2026-10-03T15:00:00.000000Z",
                  "updated_at": "2026-10-03T15:00:00.000000Z",
                  "document": {
                    "format_version": 1,
                    "nodes": [
                      {
                        "id": "callers",
                        "type": "caller_group",
                        "label": "Callers",
                        "position": {
                          "x": -80,
                          "y": 120.5
                        },
                        "capacity_rps": null,
                        "caller_count": 100,
                        "test_rps": 100
                      },
                      {
                        "id": "lb",
                        "type": "load_balancer",
                        "label": "Load balancer",
                        "position": {
                          "x": 160,
                          "y": 120.5
                        },
                        "capacity_rps": 100,
                        "routing_policy": "weighted"
                      },
                      {
                        "id": "gateway",
                        "type": "gateway",
                        "label": "Gateway",
                        "position": {
                          "x": 400,
                          "y": 120.5
                        },
                        "capacity_rps": 100,
                        "routing_policy": "round_robin"
                      },
                      {
                        "id": "server",
                        "type": "server",
                        "label": "Server",
                        "position": {
                          "x": 640,
                          "y": 120.5
                        },
                        "capacity_rps": 60
                      },
                      {
                        "id": "database",
                        "type": "database",
                        "label": "Database",
                        "position": {
                          "x": 880,
                          "y": 120.5
                        },
                        "capacity_rps": 60
                      }
                    ],
                    "edges": [
                      {
                        "id": "callers-lb",
                        "source": "callers",
                        "target": "lb",
                        "order": 0,
                        "weight": 1
                      },
                      {
                        "id": "lb-gateway",
                        "source": "lb",
                        "target": "gateway",
                        "order": 0,
                        "weight": 1.5
                      },
                      {
                        "id": "gateway-server",
                        "source": "gateway",
                        "target": "server",
                        "order": 0,
                        "weight": 0
                      },
                      {
                        "id": "server-database",
                        "source": "server",
                        "target": "database",
                        "order": 0,
                        "weight": 1
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ArchitectureNotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "503": {
            "$ref": "#/components/responses/StorageUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "put": {
        "tags": [
          "Architectures"
        ],
        "operationId": "replaceArchitecture",
        "summary": "Save a complete replacement architecture",
        "description": "Atomically replace name and entire document; no merging or upsert. Remove omitted elements and retain ID/created_at. Advance updated_at only when validated content changes. An identical PUT preserves timestamps. Incident edges must be removed with a deleted node. Unknown UUIDs return 404.",
        "requestBody": {
          "$ref": "#/components/requestBodies/ArchitectureWrite"
        },
        "responses": {
          "200": {
            "description": "Complete authoritative saved representation after commit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Architecture"
                },
                "example": {
                  "id": "0ccdc365-31f4-46c8-9f94-2b36d7f73b08",
                  "name": "Renamed application",
                  "created_at": "2026-10-03T15:00:00.000000Z",
                  "updated_at": "2026-10-03T15:01:00.000000Z",
                  "document": {
                    "format_version": 1,
                    "nodes": [
                      {
                        "id": "callers",
                        "type": "caller_group",
                        "label": "Callers",
                        "position": {
                          "x": -80,
                          "y": 120.5
                        },
                        "capacity_rps": null,
                        "caller_count": 100,
                        "test_rps": 100
                      },
                      {
                        "id": "lb",
                        "type": "load_balancer",
                        "label": "Load balancer",
                        "position": {
                          "x": 160,
                          "y": 120.5
                        },
                        "capacity_rps": 100,
                        "routing_policy": "weighted"
                      },
                      {
                        "id": "gateway",
                        "type": "gateway",
                        "label": "Gateway",
                        "position": {
                          "x": 400,
                          "y": 120.5
                        },
                        "capacity_rps": 100,
                        "routing_policy": "round_robin"
                      },
                      {
                        "id": "server",
                        "type": "server",
                        "label": "Server",
                        "position": {
                          "x": 640,
                          "y": 120.5
                        },
                        "capacity_rps": 60
                      },
                      {
                        "id": "database",
                        "type": "database",
                        "label": "Database",
                        "position": {
                          "x": 880,
                          "y": 120.5
                        },
                        "capacity_rps": 60
                      }
                    ],
                    "edges": [
                      {
                        "id": "callers-lb",
                        "source": "callers",
                        "target": "lb",
                        "order": 0,
                        "weight": 1
                      },
                      {
                        "id": "lb-gateway",
                        "source": "lb",
                        "target": "gateway",
                        "order": 0,
                        "weight": 1.5
                      },
                      {
                        "id": "gateway-server",
                        "source": "gateway",
                        "target": "server",
                        "order": 0,
                        "weight": 0
                      },
                      {
                        "id": "server-database",
                        "source": "server",
                        "target": "database",
                        "order": 0,
                        "weight": 1
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidJson"
          },
          "404": {
            "$ref": "#/components/responses/ArchitectureNotFound"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "503": {
            "$ref": "#/components/responses/StorageUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": [
          "Architectures"
        ],
        "operationId": "deleteArchitecture",
        "summary": "Delete a saved architecture",
        "description": "Atomically delete the architecture. Success has no body. Subsequent reads, writes, and deletes return 404.",
        "responses": {
          "204": {
            "description": "Architecture deleted. No response body."
          },
          "404": {
            "$ref": "#/components/responses/ArchitectureNotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "503": {
            "$ref": "#/components/responses/StorageUnavailable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/simulations": {
      "post": {
        "tags": [
          "Simulations"
        ],
        "summary": "Simulate Architecture",
        "operationId": "simulateArchitecture",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Complete recorded run. Frame zero is empty and summary equals final frame totals.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "400": {
            "description": "Structured application error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Structured application error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Structured application error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Structured application error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "description": "Calculate exactly total_ticks from the submitted document without accessing the architecture library. Return frames 0..total_ticks and the complete synchronous request/response flow. Require all components to be reachable from Caller Groups. Apply capacity only to forward requests; silently drop overflow. Responses follow their actual request path back to the original caller. Do not drain after the final tick."
      }
    }
  },
  "components": {
    "schemas": {
...
      "Architecture": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ArchitectureId",
            "readOnly": true
          },
          "name": {
            "$ref": "#/components/schemas/ArchitectureName"
          },
          "created_at": {
            "$ref": "#/components/schemas/UtcTimestamp",
            "readOnly": true
          },
          "updated_at": {
            "$ref": "#/components/schemas/UtcTimestamp",
            "readOnly": true
          },
          "document": {
            "$ref": "#/components/schemas/ArchitectureDocument"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at",
          "document"
        ]
      },
...
    },
...
    "requestBodies": {
      "ArchitectureWrite": {
        "required": true,
        "description": "Whole desired saved name/document. All typed fields required; metadata and unknown fields are forbidden.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ArchitectureWrite"
            },
            "examples": {
              "empty": {
                "summary": "Save an empty architecture",
                "value": {
                  "name": "Untitled architecture",
                  "document": {
                    "format_version": 1,
                    "nodes": [],
                    "edges": []
                  }
                }
              },
              "all_components": {
                "summary": "Save all five component types, positions, relative weights, and destination order",
                "value": {
                  "name": "Application with database",
                  "document": {
                    "format_version": 1,
                    "nodes": [
                      {
                        "id": "callers",
                        "type": "caller_group",
                        "label": "Callers",
                        "position": {
                          "x": -80,
                          "y": 120.5
                        },
                        "capacity_rps": null,
                        "caller_count": 100,
                        "test_rps": 100
                      },
                      {
                        "id": "lb",
                        "type": "load_balancer",
                        "label": "Load balancer",
                        "position": {
                          "x": 160,
                          "y": 120.5
                        },
                        "capacity_rps": 100,
                        "routing_policy": "weighted"
                      },
                      {
                        "id": "gateway",
                        "type": "gateway",
                        "label": "Gateway",
                        "position": {
                          "x": 400,
                          "y": 120.5
                        },
                        "capacity_rps": 100,
                        "routing_policy": "round_robin"
                      },
                      {
                        "id": "server",
                        "type": "server",
                        "label": "Server",
                        "position": {
                          "x": 640,
                          "y": 120.5
                        },
                        "capacity_rps": 60
                      },
                      {
                        "id": "database",
                        "type": "database",
                        "label": "Database",
                        "position": {
                          "x": 880,
                          "y": 120.5
                        },
                        "capacity_rps": 60
                      }
                    ],
                    "edges": [
                      {
                        "id": "callers-lb",
                        "source": "callers",
                        "target": "lb",
                        "order": 0,
                        "weight": 1
                      },
                      {
                        "id": "lb-gateway",
                        "source": "lb",
                        "target": "gateway",
                        "order": 0,
                        "weight": 1.5
                      },
                      {
                        "id": "gateway-server",
                        "source": "gateway",
                        "target": "server",
                        "order": 0,
                        "weight": 0
                      },
                      {
                        "id": "server-database",
                        "source": "server",
                        "target": "database",
                        "order": 0,
                        "weight": 1
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "InvalidJson": {
        "description": "The request body must contain valid JSON.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "invalid_json",
                "message": "The request body must contain valid JSON.",
                "details": []
              }
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "The request body must use application/json.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "unsupported_media_type",
                "message": "The request body must use application/json.",
                "details": []
              }
            }
          }
        }
      },
...
      "StorageUnavailable": {
        "description": "The architecture library is temporarily unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "storage_unavailable",
                "message": "The architecture library is temporarily unavailable.",
                "details": []
              }
            }
          }
        }
      },
...
    }
  }
}